--- a/core/isolate.rs
+++ b/core/isolate.rs
@@ -158,7 +158,7 @@ pub struct Isolate {
   have_unpolled_ops: bool,
   startup_script: Option<OwnedScript>,
   op_registry: OpRegistry,
-  eager_count: u32,
+  eager_poll_count: u32,
 }
 
 unsafe impl Send for Isolate {}
@@ -224,7 +224,7 @@ impl Isolate {
       pending_dyn_imports: FuturesUnordered::new(),
       startup_script,
       op_registry: OpRegistry::new(),
-      eager_count: 0,
+      eager_poll_count: 0,
     }
   }
 
@@ -317,11 +317,13 @@ impl Isolate {
       PinnedBuf::new(zero_copy_buf),
     );
 
-    let op = if isolate.eager_count != 50 {
-      isolate.eager_count += 1;
+    // To avoid latency problems we eagerly poll 50 futures and then
+    // allow to poll ops from `pending_ops`
+    let op = if isolate.eager_poll_count != 50 {
+      isolate.eager_poll_count += 1;
       match op {
         Op::Async(mut fut) => {
-          // Tries to greedily poll async ops once. Often they are immediately ready, in
+          // Tries to eagerly poll async ops once. Often they are immediately ready, in
           // which case they can be turned into a sync op before we return to V8. This
           // can save a boundary crossing.
           #[allow(clippy::match_wild_err_arm)]
@@ -334,7 +336,6 @@ impl Isolate {
         Op::Sync(buf) => Op::Sync(buf),
       }
     } else {
-      isolate.eager_count = 0;
       op
     };
 
@@ -655,6 +656,7 @@ impl Future for Isolate {
 
       // Now handle actual ops.
       self.have_unpolled_ops = false;
+      self.eager_poll_count = 0;
       #[allow(clippy::match_wild_err_arm)]
       match self.pending_ops.poll() {
         Err(_) => panic!("unexpected op error"),