@bartlomieju 2019-10-15
Ultimately we should set our eyes on switching to Tokio `0.2` as soon as it [comes out](https://gitter.im/tokio-rs/tokio?at=5da35219158cfd6735243f4d). [Looks like](https://tokio.rs/blog/2019-10-scheduler/) it's new scheduler rocks.